requirements.txt
altair==4.1.0
appdirs==1.4.3
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
boto3==1.15.11
botocore==1.18.11
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lockfile==0.12.2
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.2
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.3
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==1.0.1
pycparser==2.20
pydeck==0.5.0b1
Pygments==2.7.1
pyparsing==2.4.6
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
pyzmq==19.0.2
regex==2020.9.27
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
sentence-transformers==0.3.7.2
sentencepiece==0.1.91
six==1.14.0
streamlit==0.67.1
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.1
toolz==0.11.1
torch==1.6.0
tornado==6.0.4
tqdm==4.50.0
traitlets==5.0.4
transformers==3.3.1
tzlocal==2.1
urllib3==1.25.8
utils==1.0.1
validators==0.18.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
